+++ src/models/webview.model.ts
@@ -1,0 +1,8 @@
+import { WebviewPanel } from "vscode";
+
+export interface IWebview {
+  name: string;
+  htmlPath: string;
+  htmlContent?: string;
+  webview?: WebviewPanel;
+}
