+++ src/service/oauthService.ts
@@ -1,15 +1,12 @@
 import * as express from "express";
-import { readFileSync } from "fs";
 import { Server } from "http";
 import fetch from "node-fetch";
 import { URLSearchParams } from "url";
-import * as vscode from "vscode";
 import Commons from "../commons";
 
 export class GitHubOAuthService {
   public app: express.Express;
   public server: Server;
-  public GistSelectionView: string;
 
   constructor(
     public port: number,
@@ -18,10 +15,6 @@ export class GitHubOAuthService {
   ) {
     this.app = express();
     this.app.use(express.json(), express.urlencoded({ extended: false }));
-    this.GistSelectionView = readFileSync(
-      `${this.extensionPath}/ui/gist-selection/gist-selection.html`,
-      { encoding: "utf8" }
-    );
   }
 
   public async StartProcess() {
@@ -40,17 +33,17 @@ export class GitHubOAuthService {
       const user = await this.getUser(token);
 
       const gists = await this.getGists(token, user);
-      this.saveGistId(gists);
+      this.commons.webviewService.OpenGistSelectionpage(gists);
     });
   }
 
-  public async getToken(code: string) {
+  public getToken(code: string) {
     const params = new URLSearchParams();
     params.append("client_id", "cfd96460d8b110e2351b");
     params.append("client_secret", "ed46bd3a0f736e0da57308e86ca5fa3cf8688582");
     params.append("code", code);
 
-    return await fetch("https://github.com/login/oauth/access_token", {
+    return fetch("https://github.com/login/oauth/access_token", {
       method: "POST",
       body: params
     });
@@ -69,39 +62,6 @@ export class GitHubOAuthService {
     const currentSettings = await this.commons.GetCustomSettings();
     currentSettings.token = token;
     this.commons.SetCustomSettings(currentSettings);
-  }
-
-  public async saveGistId(gists: any) {
-    const content: string = this.GistSelectionView.replace(
-      new RegExp("@GISTS", "g"),
-      JSON.stringify(gists)
-    ).replace(
-      new RegExp("@PWD", "g"),
-      vscode.Uri.file(this.extensionPath)
-        .with({
-          scheme: "vscode-resource"
-        })
-        .toString()
-    );
-    const gistSelectionPanel = vscode.window.createWebviewPanel(
-      "selectGist",
-      "Select Your Existing Gist",
-      vscode.ViewColumn.One,
-      {
-        retainContextWhenHidden: true,
-        enableScripts: true
-      }
-    );
-    gistSelectionPanel.webview.html = content;
-    gistSelectionPanel.webview.onDidReceiveMessage(async message => {
-      if (!message.close) {
-        const extSettings = await this.commons.GetSettings();
-        extSettings.gist = message.id;
-        this.commons.SaveSettings(extSettings);
-      } else {
-        gistSelectionPanel.dispose();
-      }
-    });
   }
 
   public async getUser(token: string) {
